Flask-Docker/app/api/requirements.txt
flask>=2.1.3
pandas==1.4.3
scikit-learn==1.1.1
gevent==22.10.2
Pyrebase4==4.5.0